==> notebooks/p02_creating_truth_datasets_for_testing_simba.ipynb <==
In this notebook we create several truth datasets for testing the precision and recall of SIMBA from several existing publically available databases/studies. 

### Loading libraries

%matplotlib inline
import matplotlib
import matplotlib.pyplot as plt
plt.rcParams['figure.dpi'] = 150
plt.rcParams['savefig.dpi'] = 200
plt.rcParams['pdf.fonttype'] = 42
plt.rcParams['ps.fonttype'] = 42

params = {'legend.fontsize': 'small',
         'axes.labelsize': 'small',
         'axes.titlesize':'small',
         'xtick.labelsize':'small',
         'ytick.labelsize':'small',
         'figure.figsize': (3, 2)}
plt.rcParams.update(params)
import os
import glob
import re
import itertools
import collections
from collections import Counter
import math
import random
from random import randrange
import string
import subprocess
import numpy as np
import pandas as pd
from pandas import ExcelWriter
from pandas import ExcelFile
import seaborn as sns 
import phylopandas as ph
cmap = sns.cm.rocket_r
sns.set_style("white")
import anndata
from anndata import read_h5ad
from anndata import AnnData
import phylopandas as ph
import wget


### Directories

mainDir = '/oak/stanford/groups/quake/gita/raw/nb/microbe/paper/forGitHub/post_review/'
mainDir1= mainDir + 'human_Gastrointestinal_Bacteria_Culture_Collection/'
mainDir2='/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/bacterial_test/batch_of_ten/'
mainDir3 = '/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/bacterial_test/refseq/'
mainDir4 = '/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/fungal/'
tosave = mainDir + 'files_to_upload_to_github'
dbDir = '/oak/stanford/groups/quake/gita/raw/database/'


### One batch of 500 random bacterial seqs from the human gut

first going to read a tsv file (filereport_read_run_PRJEB23845_tsv.txt downloaded from https://www.ebi.ac.uk/ena/browser/view/PRJEB23845?show=reads) containing the ftp addresses for genomes in the human gut microbiome isolates so I can download them and randomly select regions from various genomes to build a truth dataset for testing how well SIMBA can detect microbes. 

The associated paper:
https://www.nature.com/articles/s41587-018-0009-7#data-availability
Human Gastrointestinal Bacteria Culture Collection


cd $mainDir1

addresses = pd.read_csv(mainDir + 'filereport_read_run_PRJEB23845_tsv.txt', delimiter='\t')

got the ftp addresses

mainDir = '/oak/stanford/groups/quake/gita/raw/nb/microbe/paper/forGitHub/post_review/'
mainDir1= mainDir + 'human_Gastrointestinal_Bacteria_Culture_Collection/'
mainDir2='/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/bacterial_test/batch_of_ten/'
mainDir3 = '/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/bacterial_test/refseq/'
mainDir4 = '/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/fungal/'
tosave = mainDir + 'files_to_upload_to_github'

dbDir = '/oak/stanford/groups/quake/gita/raw/database/'


addresses.head(3)

first getting unique taxids from this study

taxids_unique=list(addresses.tax_id.unique())

we can use the taxids to retrieve ftp paths for downloading genomes from this assembly report downloaded from NCBI (under 16): https://www.ncbi.nlm.nih.gov/genome/doc/ftpfaq/ 

wget.download('https://ftp.ncbi.nlm.nih.gov/genomes/refseq/bacteria/assembly_summary.txt')

now getting the files
first removed the comment line (first line from the file using the following command in the command line: sed -i '1d' assembly_summary.txt)

assembly_report=pd.read_csv(mainDir1 + 'assembly_summary.txt', delimiter='	')
header = ['assembly_accession', 'bioproject', 'biosample', 'wgs_master',
       'refseq_category', 'taxid', 'species_taxid', 'organism_name',
       'infraspecific_name', 'isolate', 'version_status', 'assembly_level',
       'release_type', 'genome_rep', 'seq_rel_date', 'asm_name', 'submitter',
       'gbrs_paired_asm', 'paired_asm_comp', 'ftp_path',
       'excluded_from_refseq', 'relation_to_type_material',
       'asm_not_live_date']
assembly_report.columns = header
assembly_report.head(3)

now getting the assembly report dataframe to only contain species from the gut study

assembly_report_fil = assembly_report[assembly_report['taxid'].isin(taxids_unique)]
assembly_report_fil.shape


taking only unique entries

assembly_report_fil2 = assembly_report_fil.drop_duplicates(subset=['taxid'])
#just renaming the dataframe for easier access
arf = assembly_report_fil2 

the ftp path doesn't take me to the genome. It takes me to a folder that contains the genome, so I appended the right address to be able to get the genome

arf['genome']= arf['ftp_path'].apply(lambda x: x.split('/')[-1] + '_genomic.fna.gz')
arf['genome_link'] = arf['ftp_path'] + '/' + arf['genome']
ftps = assembly_report_fil2['genome_link'].tolist()

downloading the genomes

for path in ftps:
    wget.download(path)

Now I want to get 10 random sequences, each 100 bases long, from 50 random bacterial genomes (selected from the gut) and collect them as part of one dataframe. Once I have the dataframe, I will use phylopandas to write a fasta file. This will become the test bacterial dataset. Will test it before adding viruses & fungi

#sampling x sequences from each genome
x=10 #number of sequences from a genome to obtain
seq_len=100 #length of the sequence to blast
num_genomes = 50


random_seq_df=pd.DataFrame({})
counter =0
for file in glob.glob(mainDir1 + '*.fna'):  
    df = ph.read_fasta(file)
    if counter<num_genomes:
        if df.shape[0]>=x: #if the genome file has at least 10 long seqs associated with it to choose random seqs from
            counter=counter+1 #want only to sample 50 genomes
            genome=file.split('/')[-1] + '.gz'
            sub_df = df.iloc[:x, :]
            sub_df['rand_seq'] = sub_df['sequence'].apply(lambda x: x[100:100+seq_len]) #start at 500bp for each contig and take 100 bases
            sub_df['genome']=genome
            random_seq_df=pd.concat([random_seq_df, sub_df])    
    else:
        break
    

#merging dataframes to get information about each random seq 
arf_sub=arf[['assembly_accession','bioproject','taxid','organism_name','infraspecific_name','ftp_path','genome']]
random_seq_df_sub=random_seq_df[['id','rand_seq','description','genome']]
blast_df=random_seq_df_sub.merge(arf_sub, on='genome')
blast_df.shape #should be 500 (50 genomes, 10 random segments each)

blast_df.head(3)

saving the resulting dataframe and then saving it in fasta form as well 

blast_df[['taxid', 'organism_name','infraspecific_name', 'genome', 'ftp_path', 'assembly_accession'
         ]].to_csv('/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/bacterial_test/random_contigs_gut_bacterial_genomes_500_contigs.csv')

This will write a file called "random_contigs_gut_bacterial_genomes_500_contigs.fasta".  
The headers follow this format: **'taxon id'_'ncbi_contig_id'_'organism_name'**:

89014_NZ_JAAITI010000096.1_Blautia luti


# #Opening a fasta file to write random sequences to
f= open('random_contigs_gut_bacterial_genomes_500_contigs.fasta', 'w')
for i in list(range(0,blast_df.shape[0])):
    header = str(blast_df['taxid'].iat[i]) + '_'+ str(blast_df['id'].iat[i]) + '_' + blast_df['organism_name'].iat[i]
    sequence = blast_df['rand_seq'].iat[i]
    f.write('>' + header + '\n' + sequence + '\n')
f.close()


### 10 batches of 500 random bacterial seqs from the human gut (total 5000 seqs)
still from 50 random species (100 regions instead of 10). There is a possibility that 100 regions will enable higher precision/recall than 10 regions. 

!mkdir /oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/bacterial_test/batch_of_ten

#sampling x sequences from each genome
x=10 #number of sequences from a genome to obtain
seq_len=100 #length of the sequence to blast
num_genomes = 50


random_seq_df=pd.DataFrame({})
counter =0
for file in glob.glob(mainDir1 + '*.fna'):  
    df = ph.read_fasta(file)
    if (counter<num_genomes): #if the genome file has at least x long seqs associated with it to choose random seqs from & #making sure the shortest contig is at least 6000 bases 
        if df.shape[0]>=x: 
            df['contig_length'] = df.sequence.apply(lambda x: len(x)) #getting the length of each contig 
            df = df.sort_values(by='contig_length', ascending=False) #sorting so that I select long contigs so that I can select multiple seqs from each contig
            if df['contig_length'].iat[x]> 6000:
                counter=counter+1 #want only to sample 50 genomes
                genome=file.split('/')[-1] + '.gz'
                sub_df = df.iloc[:x, :]            
        
            for i in range(100,5100, 400): #taking 10 random segments from a contig, starting at 100 base position
                sub_df['rand_seq'] = sub_df['sequence'].apply(lambda x: x[i:i+seq_len]) 
                sub_df['genome']=genome
                random_seq_df=pd.concat([random_seq_df, sub_df]) 
    else:
        break
    

#merging dataframes to get information about each random seq 
arf_sub=arf[['assembly_accession','bioproject','taxid','organism_name','infraspecific_name','ftp_path','genome']]
random_seq_df_sub=random_seq_df[['id','rand_seq','description','genome']]
blast_df=random_seq_df_sub.merge(arf_sub, on='genome')
blast_df.shape 

dropping any potential duplicates and sequences that have "N"

blast_df = blast_df.drop_duplicates()
blast_df = blast_df[~blast_df['rand_seq'].str.contains('N')]

selecting only 100 contigs per genome (currently there are about 130)

blast_df_100 = pd.DataFrame({})
for genome in blast_df.genome.unique():
    sub = blast_df[blast_df['genome']==genome].sample(n=100) 
    blast_df_100=pd.concat([blast_df_100, sub])

will create a unique id column (id2) otherwise segments from the same contig will have the same id in the resulting fasta file. 

id2=[str(x) for x in list(range(blast_df_100.shape[0]))]
blast_df_100['id2']=id2
blast_df_100['id2'] = blast_df_100['id'] + '_' + blast_df_100['id2']


spliting the main CSV file so I can run 10 SIMBA jobs in parallel. The following will produce both csvs and fasta files (500 sequences per file)

readsDir = mainDir2 + 'fasta_files/'


!mkdir $readsDir

num_seq_per_batch=500 #how many seqs per fasta file
for i in range(0,blast_df_100.shape[0],num_seq_per_batch):
    start=i
    end=i+ num_seq_per_batch
    batch_df = blast_df_100.iloc[start:end]
    batch_df.to_csv(readsDir + 'batch_' + str(i) + '.csv')
    
    # #Opening a fasta file to write random sequences to
    f= open(readsDir + 'batch_' + str(i) + '.fasta', 'w')
    for i in list(range(0,batch_df.shape[0])):
        header = str(batch_df['taxid'].iat[i]) + '_'+ str(batch_df['id2'].iat[i]) + '_' + batch_df['organism_name'].iat[i]
        header = header.replace(' ', '_')
        sequence = batch_df['rand_seq'].iat[i] 
        f.write('>' + header + '\n' + sequence + '\n')
    f.close()



the above code works now, but there was an error at the level of fasta file writing wherer it was inputing the same sequence. I fixed it, and use the below block to write new files. The csvs were written correctly

num_seq_per_batch=500 #how many seqs per fasta file
for file in glob.glob('/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/bacterial_test/batch_of_ten/fasta_files/*.csv'):
    batch_df = pd.read_csv(file)
    filename = file.split('/')[-1].split('.csv')[0]
    #Opening a fasta file to write random sequences to
    f= open(readsDir + filename + '.fasta', 'w')
    for i in list(range(0,batch_df.shape[0])):
        header = str(batch_df['taxid'].iat[i]) + '_'+ str(batch_df['id2'].iat[i]) + '_' + batch_df['organism_name'].iat[i]
        header = header.replace(' ', '_')
        sequence = batch_df['rand_seq'].iat[i]
        f.write('>' + header + '\n' + sequence + '\n')
    f.close()


### RefSeq bacterial truth dataset (100 seqs per genome, 10 batches for a total of 5000 seqs)
I want to see how SIMBA does in picking up bacteria in general, as opposed to bacterial species found in the gut. So here I will select 100 contigs per genome from 50 randomly selected Refseq bacterial species. 

no need to rerun this, as I already have refseq database saved as a dataframe. See databaseCreationSC_microbeDB_fungalDB_jan2021


refseq_dir = '/oak/stanford/groups/quake/gita/raw/tab2_20200508/tab2microbial/'


#Reading in the original fasta file, quickly converting it to dataframe
init=pd.DataFrame()
for file in glob.glob(dbDir + 'bactRepSeqs_Dec2020/ncbi-genomes-2020-12-16/*.fna'):
    df1 = ph.read_fasta(file)
    init = pd.concat([init, df1])

    
#extracting only genome names by replacing the genbank id
init['name']=init['description'].str.replace('N\w_\w+\d+.\d+ ', '')

#changing the names to contain only species information (don't want strains)
init['sp'] =init.name.str.split('strain')
init['sp'] = init['sp'].apply(lambda x: x[0])
init['sp2'] = init['sp'].apply(lambda x: x.split(' ')[0:2])
init['sp2'] = init['sp2'].apply(lambda x: ' '.join(x))

init.to_csv(refseq_dir + 'refSeqBacterial_Dec2020.csv')

reading the refseq database (from Dec 2020) and dropping duplicate species (i.e. different strains of the same species)


refseq =pd.read_csv(refseq_dir + 'refSeqBacterial_Dec2020.csv')
print('number of unique species in the refseq db:', refseq['sp2'].nunique())
print('total number of species in the refseq db:', len(refseq['sp2']))

refseq_unique=refseq.drop_duplicates(subset='name', keep='first')
print('number of unique species in refseq:', refseq_unique.shape)
refseq_unique.head(1)


first I want to randomly select 50 species. first I will select those species for which I can find taxon id by matching their names to NCBI's assembly dataframe.  

orgs_with_taxid=list(set(refseq_unique['sp2']).intersection(set(assembly_report['organism_name']))) #organism with taxid in the refseq
refseq_unique = refseq_unique[refseq_unique['sp2'].isin(orgs_with_taxid)] #filtering refseq to contain only those organisms 
refseq_unique['genome_length'] = refseq_unique['sequence'].apply(lambda x: len(x)) #getting genome length too
refseq_unique_fil=refseq_unique[refseq_unique['genome_length']> 1e5] #making sure the genome is long enough to select 100 random regions without the regions being too close to each other
ref50 = refseq_unique_fil.sample(n=50)[['id', 'sequence', 'name', 'sp2', 'genome_length']] #randomly selecting 50 species




selecting 100 segments per genome (50 genomes)


#selecting 100 segments per genome (50 genomes)
#sampling x sequences from each genome
x=100 #number of sequences from a genome to obtain
seq_len=100 #length of the sequence to blast
num_genomes = 50

ref50_randseqs=pd.DataFrame({})
randseqs =[]
counter = 0
for index in range(ref50.shape[0]):
    genome_length = ref50['genome_length'].iat[index]
    for i in list(range(seq_len,genome_length-seq_len, 1000))[:x]: #taking 100 random segments from a genome
        ref50_randseqs=pd.concat([ref50_randseqs, pd.DataFrame(ref50.iloc[index]).T])
        randseqs.append(ref50['sequence'].iat[index][i:i+seq_len])
        counter=counter + 1
    
ref50_randseqs['rand_seq']= randseqs

ref50_randseqs.shape





ensuring there are no duplicate sequences or ones containing "N"

ref50_randseqs = ref50_randseqs.drop_duplicates()
ref50_randseqs = ref50_randseqs[~ref50_randseqs['rand_seq'].str.contains('N')]
ref50_randseqs.shape

adding a unique id column (id2) since random sequences from the same genome currently have the same id. 

id2=[str(x) for x in list(range(ref50_randseqs.shape[0]))]
ref50_randseqs['id2']= id2
ref50_randseqs['id2'] = ref50_randseqs['id'] + '_' + ref50_randseqs['id2']
ref50_randseqs = ref50_randseqs.reset_index().drop(columns={'index'}) #reset index

ref50_randseqs.head(1)

adding taxids by getting them from ncbi assembly report

assembly_report_short = assembly_report[['taxid', 'organism_name','infraspecific_name', 'ftp_path']]
assembly_report_short = assembly_report_short.rename(columns={'organism_name':'sp2'})
a = ref50_randseqs['sp2'].unique().tolist()
b = assembly_report_short[assembly_report_short['sp2'].isin(a)].drop_duplicates('sp2')
ref50_randseqs_withtaxid = ref50_randseqs.merge(b, on='sp2', how='left')

in addition to taxids I want to add taxonomic lineages 

tax = pd.read_csv(dbDir + 'taxonomyNCBI20200125/ncbi_lineages_2021-01-26.csv')
#want to take only the following columns from the lineage dataframe tax 
tax_short=tax[['tax_id','superkingdom', 'phylum', 'class', 'order', 'family', 'genus', 'species']]
tax_short = tax_short.rename(columns={'tax_id':'taxid'})

ref50_lineage=ref50_randseqs_withtaxid.merge(tax_short, on='taxid', how='left')
ref50_lineage.head(1)

now dividing this dataframe into 10 csvs and fasta files to run 10 SIMBA jobs in parallel

readsDir2 = mainDir3 + 'fasta_files/'


!mkdir $readsDir2

num_seq_per_batch=500 #how many seqs per fasta file
for i in range(0,ref50_lineage.shape[0],num_seq_per_batch):
    start=i
    end=i+ num_seq_per_batch
    batch_df = ref50_lineage.iloc[start:end]
    batch_df.to_csv(readsDir2 + 'batch_' + str(i) + '.csv')
    
    #Opening a fasta file to write random sequences to
    f= open(readsDir2 + 'batch_' + str(i) + '.fasta', 'w')
    for i in list(range(0,batch_df.shape[0])):
        header = str(batch_df['taxid'].iat[i]) + '_'+ str(batch_df['id2'].iat[i]) + '_' + batch_df['sp2'].iat[i]
        header = header.replace(' ', '_')
        sequence = batch_df['rand_seq'].iat[i]
        f.write('>' + header + '\n' + sequence + '\n')
    f.close()


### Human only dataset 
I want to see what are the rates of false positives and negatives when we only have 2400 sequences randomly selected from the human genome (100 seqs per chromosomes 1-22 and X, Y). I have the human genome assembly from GCF_000001405.39_GRCh38.p13_genomic.fna 

humanGenome = dbDir + 'human_genome_and_synthetic_chromosomes_july2021/GCF_000001405.39_GRCh38.p13_genomic.fna'

reading the human genome fasta file as a csv since it's easier to select from it when its a datafame

hum=ph.read_fasta(humanGenome)

hum.head(3)

seq_len=100
x=100 #how many sequences per chromosome to take 
hum_randseqs=pd.DataFrame({})
randseqs =[]
chromosomes=[str(x+1) for x in list(range(22))]
chromosomes.append('X')
chromosomes.append('Y')

for i in chromosomes: 
    sub = hum[hum['description'].str.contains('Homo sapiens chromosome ' + i) ].iloc[1,:] #taking one of the segments from each chromosome to randomly select 100 sequences from
    sub['chromosome']= i
    segment_length = len(sub['sequence'])
    for j in list(range(seq_len+1000, segment_length-seq_len, 300))[:x]: #taking 100 random segments from each chromosome
        hum_randseqs=pd.concat([hum_randseqs, pd.DataFrame(sub).T])
        randseqs.append(sub['sequence'][j:j+seq_len])

hum_randseqs['rand_seq']= randseqs

id2=[str(x) for x in list(range(hum_randseqs.shape[0]))]
hum_randseqs['id2']=id2
hum_randseqs['id2'] = hum_randseqs['chromosome'] + '_' + hum_randseqs['id'] + '_' + hum_randseqs['id2']
hum_randseqs = hum_randseqs.reset_index().drop(columns={'index'}) #reset index


hum_randseqs.head(2)

making sure there are not any sequences with "N"

hum_randseqs[~hum_randseqs['rand_seq'].str.contains('N')].shape[0]

Let's add some unique IDs before saving in batches (for parallel jobs) both as csv and fasta file 

!mkdir /oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/human_test/fasta_files

also saving the data in batches (for parallel jobs) and saving it to csv & fasta

humdir = '/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/human_test/fasta_files/'
num_seq_per_batch=500 #how many seqs per fasta file
for i in range(0,hum_randseqs.shape[0], num_seq_per_batch):
    start=i
    end=i+ num_seq_per_batch
    batch_df = hum_randseqs.iloc[start:end]
    batch_df.to_csv(humdir + 'batch_' + str(i) + '.csv')

    # Opening a fasta file to write batch to 
    f= open(humdir + 'batch_' + str(i) + '.fasta', 'w')
    for j in list(range(0,batch_df.shape[0])):
        header = batch_df['id2'].iat[j] 
        sequence = batch_df['rand_seq'].iat[j]
        f.write('>' + header + '\n' + sequence + '\n')
    f.close()


### Viral (1000 total sequences, 100 genomes, 10 contigs per genome)
Getting test viral sequences
Since I already have a fasta file corresponding to the viral refseq database, I will subsample it. Downloaded March, 2020

virus_df=ph.read_fasta('/oak/stanford/groups/quake/gita/raw/database/viral_nr_march25_2020/viral_refseq_nuc/all_refseq_viral_genomes.fasta')

virus_df.head(2)

virus_df.id.nunique()

randomly sampling 100 viruses, and taking 10 random segments (100 bases long) from each genome.

virus_df['genome_length']= virus_df['sequence'].apply(lambda x: len(x))
virus_df_100 = virus_df[virus_df['genome_length']>5000].sample(100) #making sure genomes are long enough to be selected for this exercise

seq_len=100
x=10 #how many sequences per genome to take 
vir_randseqs=pd.DataFrame({})
randseqs =[]

for i in range(virus_df_100.shape[0]): 
    sub = virus_df_100.iloc[i]
    genome_length = sub['genome_length']
    for j in list(range(seq_len+300, genome_length-seq_len, 300))[:x]: #taking 10 random segments from each genome
        vir_randseqs=pd.concat([vir_randseqs, pd.DataFrame(sub).T])
        randseqs.append(sub['sequence'][j:j+seq_len])

vir_randseqs['rand_seq']= randseqs

id2=[str(x) for x in list(range(vir_randseqs.shape[0]))]
vir_randseqs['id2']=id2
vir_randseqs['id2'] =  vir_randseqs['id'] + '_' + vir_randseqs['id2']
vir_randseqs['description'] = vir_randseqs['description'].apply(lambda x: x.replace(' ', '_'))
vir_randseqs = vir_randseqs.reset_index().drop(columns={'index'}) #reset index


writing fasta and csv files from the dataframe to input into SIMBA for testing (10 jobs, 100 seqs each)

virdir = '/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/viral_test/fasta_files/'
num_seq_per_batch=100 #how many seqs per fasta file
for i in range(0,vir_randseqs.shape[0], num_seq_per_batch):
    start=i
    end=i+ num_seq_per_batch
    batch_df = vir_randseqs.iloc[start:end]
    batch_df.to_csv(virdir + 'batch_' + str(i) + '.csv')

    # Opening a fasta file to write batch to 
    f= open(virdir + 'batch_' + str(i) + '.fasta', 'w')
    for j in list(range(0,batch_df.shape[0])):
        header = batch_df['id2'].iat[j] + '_' + batch_df['description'].iat[j]
        sequence = batch_df['rand_seq'].iat[j]
        f.write('>' + header + '\n' + sequence + '\n')
    f.close()


### eHOMD (expanded Human Oral Microbiome Dataset)
Besides the bacterial refseq and the gut bacterial truth datasets, I want to create one that has bacteria from differrent parts of the body, namely several sites from the oral microbiome. First I downloaded a manifest file from HOMD, which I downloaded to OAK using the followinig command from my local computer 
rsync ~/Downloads/HOMD_taxon_table2022-09-15_1663245201.txt  sherlock://oak/stanford/groups/quake/gita/raw/nb/microbe/paper/forGitHub/post_review/homd/.

I also got other info related to each species from a seperate file: HOMD_genome_table2022-09-15_1663245202.txt

Both files were downloaded from: https://www.homd.org/download#genome

homd_dir = '/oak/stanford/groups/quake/gita/raw/nb/microbe/paper/forGitHub/post_review/homd/'
homd_file = 'HOMD_taxon_table2022-09-15_1663245201.txt'
homd=pd.read_csv(homd_dir + homd_file, delimiter='	')
homd.columns = [x.lower() for x in homd.columns.tolist()] #making sure the column headers are lower case for easier typing
homd.head()

reading the meta file and selecting 200 genomes ...

homd_sub=homd[homd['body_site']=='Oral']

homd_sub2 = homd_sub.sample(n=200)
homd_sub3 = homd_sub2.drop_duplicates(['ncbi_taxon_id'])

taxa = homd_sub3['ncbi_taxon_id'].unique().tolist()

saving homd_sub3 because it is a random selection and I don't want to have it changeding the file

homd_sub3.to_csv(homd_dir + 'homd_200_genomes_randomly_selected.csv')

meta_file = homd_dir + 'HOMD_genome_table2022-09-15_1663245202.txt'
meta = pd.read_csv(meta_file, delimiter='	')
meta.columns

reading NCBI's assembly report to be able to download the genomes in the homd data based on their Genbank assmebly numbers

assembly_report=pd.read_csv(mainDir1 + 'assembly_summary.txt', delimiter='	')
header = ['assembly_accession', 'bioproject', 'biosample', 'wgs_master',
       'refseq_category', 'taxid', 'species_taxid', 'organism_name',
       'infraspecific_name', 'isolate', 'version_status', 'assembly_level',
       'release_type', 'genome_rep', 'seq_rel_date', 'asm_name', 'submitter',
       'gbrs_paired_asm', 'paired_asm_comp', 'ftp_path',
       'excluded_from_refseq', 'relation_to_type_material',
       'asm_not_live_date']
assembly_report.columns = header
assembly_report.head(3)

assembly_report_fil = assembly_report[assembly_report['taxid'].isin(taxa)]
assembly_report_fil.shape


I only need 50 genomes, so I will drop any duplicate taxids and take the first 50

arf2 =assembly_report_fil.drop_duplicates(['taxid'])[:50]

downloading the genomes

download_dir = homd_dir + 'genomes_from_homd_50'
!cd $download_dir

as before the ftp_path doesn't take me to the genome, so need to get the genomic path

arf2['genome']= arf2['ftp_path'].apply(lambda x: x.split('/')[-1] + '_genomic.fna.gz')
arf2['genome_link'] = arf2['ftp_path'] + '/' + arf2['genome']
ftps = arf2['genome_link'].tolist()

I got these files by creating a list in the terminal and running the following in the terminal

lst=('path1 path2 ... pathn')
for f in $lst; do  wget $f; done

putting together genomic fasta files that were downloaded into dataframe format for easier processing and sampling. 

homd_reads =pd.DataFrame({})
for file in glob.glob(homd_dir + 'genomes_from_homd_50/*.fna'):
    df = ph.read_fasta(file)
    df['filename']= file.split('/')[-1].split('_genomic.fna')[0]
    df=df[['id', 'sequence', 'description', 'filename']]
    homd_reads = pd.concat([homd_reads, df])


homd_reads.head(1)

now I will do everything I did for getting gut bacterial sequences (5000 seqs) to be processed as batch of 10 fasta files for being input to SIMBA

#sampling x sequences from each genome
x=100 #number of sequences from a genome to obtain
seq_len=100 #length of the sequence to blast
num_genomes = 50

homd_reads['genome_length'] = homd_reads['sequence'].apply(lambda x: len(x))
ref50=homd_reads[homd_reads['genome_length']>=50000] #taking only genomic segments that are 50k long 
ref50 = ref50.groupby('filename').apply(lambda x: x.sample(1)).reset_index(drop=True) 

ref50_randseqs=pd.DataFrame({})
randseqs =[]
counter = 0
for index in range(ref50.shape[0]):
    genome_length = ref50['genome_length'].iat[index]
    for i in list(range(seq_len,genome_length-seq_len, 500))[:x]: #taking 100 random segments from a genome
        ref50_randseqs=pd.concat([ref50_randseqs, pd.DataFrame(ref50.iloc[index]).T])
        randseqs.append(ref50['sequence'].iat[index][i:i+seq_len])
        counter=counter + 1
    
ref50_randseqs['rand_seq']= randseqs

print(ref50_randseqs.shape[0])

# ensuring there are no duplicate sequences or ones containing "N"
ref50_randseqs = ref50_randseqs.drop_duplicates()
ref50_randseqs = ref50_randseqs[~ref50_randseqs['rand_seq'].str.contains('N')]
print(ref50_randseqs.shape[0])

# adding a unique id column (id2) since random sequences from the same genome currently have the same id. 
id2=[str(x) for x in list(range(ref50_randseqs.shape[0]))]
ref50_randseqs['id2']= id2
ref50_randseqs['id2'] = ref50_randseqs['id'] + '_' + ref50_randseqs['id2']
ref50_randseqs = ref50_randseqs.reset_index().drop(columns={'index'}) #reset index


ref50_randseqs.head(1)

#adding taxids by getting them from ncbi assembly report
arf3 = arf2[['taxid', 'organism_name','infraspecific_name', 'ftp_path', 'assembly_accession']]
ref50_randseqs['assembly_accession'] = ref50_randseqs['filename'].apply(lambda x: '_'.join(x.split('_')[0:2]))
ref50_randseqs_withtaxid = arf3.merge(ref50_randseqs, on='assembly_accession', how='right')

#in addition to taxids I want to add taxonomic lineages 
tax = pd.read_csv(dbDir + 'taxonomyNCBI20200125/ncbi_lineages_2021-01-26.csv')
#want to take only the following columns from the lineage dataframe tax 
tax_short=tax[['tax_id','superkingdom', 'phylum', 'class', 'order', 'family', 'genus', 'species']]
tax_short = tax_short.rename(columns={'tax_id':'taxid'})

ref50_lineage=ref50_randseqs_withtaxid.merge(tax_short, on='taxid', how='left')
ref50_lineage['sp2']=ref50_lineage['species'].apply(lambda x: "_".join(x.split(' ')).lower())
ref50_lineage.head(1)


#now diving this dataframe into 10 csvs and fasta files to run 10 SIMBA jobs in parallel
readsDir2 = '/oak/stanford/groups/quake/gita/raw/tab3-14_20210420/benchmarking_post_reviews/bacterial_test/homd/fasta_files/'

num_seq_per_batch=500 #how many seqs per fasta file
for i in range(0,ref50_lineage.shape[0],num_seq_per_batch):
    start=i
    end=i+ num_seq_per_batch
    batch_df = ref50_lineage.iloc[start:end]
    batch_df.to_csv(readsDir2 + 'batch_' + str(i) + '.csv')
    
    #Opening a fasta file to write random sequences to
    f= open(readsDir2 + 'batch_' + str(i) + '.fasta', 'w')
    for i in list(range(0,batch_df.shape[0])):
        header = str(batch_df['taxid'].iat[i]) + '_'+ str(batch_df['id2'].iat[i]) + '_' + batch_df['sp2'].iat[i]
        header = header.replace(' ', '_')
        sequence = batch_df['rand_seq'].iat[i]
        f.write('>' + header + '\n' + sequence + '\n')
    f.close()

### Creating a fungal dataset

first downloading the NCBI genome assembly report for eukaryotes. After downloadinig NCBI's "dataset" package I was able to download the genomes using a list of assembly accession numbers. Here's the command I ran in the terminal (after activating the mainEnv2 environment). There is also a python api, but it was not working for me. Might try again later. see documentation here https://www.ncbi.nlm.nih.gov/datasets/docs/v1/download-and-install/

note list of accessions came from a study of fungi. https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-017-0373-4#Sec20


conda install -c conda-forge ncbi-datasets-cli"<14

acc=('GCA_000002945.2 GCA_000002515.1 GCA_001708105.1 GCA_011022315.1 GCA_003290485.1 GCA_000710275.1 GCA_001190045.1 GCA_001642055.1 GCA_000769745.1 GCA_016906535.1 GCA_000313485.1 GCA_002922495.1 GCA_003708225.2 GCA_015533775.1 GCA_008728235.1 GCA_000835515.1 GCA_016952355.1 GCA_001263375.1 GCA_001278385.1 GCA_001447935.2 GCA_001481775.2 GCA_022641835.1 GCA_900080155.1 GCA_001936155.1 GCA_001984395.2 GCA_022343325.1 GCA_004918485.1 GCA_900322975.1 GCA_018344385.1 GCA_003856795.1 GCA_001736105.1 GCA_900149155.1 GCA_016952305.1 GCA_902749255.1 GCA_016906865.1 GCA_001675385.1 GCA_016767815.1 GCA_009663055.1 GCA_900002835.2 GCA_004920355.1 GCA_001854445.2 GCA_019321675.1 GCA_010883545.1 GCA_008692325.1 GCA_015344875.1 GCA_943193715.1 GCA_020647795.1 GCA_010232095.1 GCA_018140865.1 GCA_023620035.1')

for f in $acc;do datasets download genome accession $f --filename $f.zip;done


getting the fungi assembly accessions from this document from ncbi. Will capture the genomes that it shares in common with the study linked above

wget.download('https://ftp.ncbi.nlm.nih.gov/genomes/GENOME_REPORTS/eukaryotes.txt')

moving it to the fungal directory

mv $mainDir/eukaryotes.txt $mainDir4

euks_assembly = pd.read_csv(mainDir4 + 'eukaryotes.txt', delimiter='\t')
euks_assembly = euks_assembly.rename(columns={'Organism/Name':'organism','TaxID':'taxid'})

euks_assembly.head(1)

found a list of fungal species in the gut

fung=pd.read_excel(mainDir4 + '40168_2017_373_MOESM3_ESM.xlsx').sort_values(by='Reads (n=27,091,491,028)', ascending=False).iloc[1:100]
fung_as=euks_assembly[euks_assembly['organism'].isin(fung['Name'].tolist())].drop_duplicates('organism')[:50]
fung_as.head(1)

as before loading the data as a dataframe for easier processing

fung_reads =pd.DataFrame({})
for file in glob.glob(mainDir4 + 'genomes/*.fna'):
    df = ph.read_fasta(file)
    df['filename']= file.split('/')[-1].split('_genomic.fna')[0]
    df=df[['id', 'sequence', 'description', 'filename']]
    fung_reads = pd.concat([fung_reads, df])

fung_reads['Assembly Accession'] = fung_reads['filename'].apply(lambda x: '_'.join(x.split('_')[:-1]))

now I will do everything I did for getting gut bacterial sequences (5000 seqs) to be processed as batch of 10 fasta files for being input to SIMBA

#sampling x sequences from each genome
x=100 #number of sequences from a genome to obtain
seq_len=100 #length of the sequence to blast
num_genomes = 50

fung_reads['genome_length'] = fung_reads['sequence'].apply(lambda x: len(x))
ref50=fung_reads[fung_reads['genome_length']>=50000] #taking only genomic segments that are 50k long 
ref50 = ref50.groupby('filename').apply(lambda x: x.sample(1)).reset_index(drop=True) 

ref50_randseqs=pd.DataFrame({})
randseqs =[]
counter = 0
for index in range(ref50.shape[0]):
    genome_length = ref50['genome_length'].iat[index]
    for i in list(range(seq_len,genome_length-seq_len, 500))[:x]: #taking 100 random segments from a genome
        ref50_randseqs=pd.concat([ref50_randseqs, pd.DataFrame(ref50.iloc[index]).T])
        randseqs.append(ref50['sequence'].iat[index][i:i+seq_len])
        counter=counter + 1
    
ref50_randseqs['rand_seq']= randseqs

print(ref50_randseqs.shape[0])

# ensuring there are no duplicate sequences or ones containing "N"
ref50_randseqs = ref50_randseqs.drop_duplicates()
ref50_randseqs = ref50_randseqs[~ref50_randseqs['rand_seq'].str.contains('N')]
print(ref50_randseqs.shape[0])

# adding a unique id column (id2) since random sequences from the same genome currently have the same id. 
id2=[str(x) for x in list(range(ref50_randseqs.shape[0]))]
ref50_randseqs['id2']= id2
ref50_randseqs['id2'] = ref50_randseqs['id'] + '_' + ref50_randseqs['id2']
ref50_randseqs = ref50_randseqs.reset_index().drop(columns={'index'}) #reset index


tax = pd.read_csv(dbDir + 'taxonomyNCBI20200125/ncbi_lineages_2021-01-26.csv')
#want to take only the following columns from the lineage dataframe tax 
tax_short=tax[['tax_id','superkingdom', 'phylum', 'class', 'order', 'family', 'genus', 'species']]
tax_short = tax_short.rename(columns={'tax_id':'taxid'})

ref50_lineage = ref50_randseqs.merge(tax_short, on='taxid', how='left')

ref50_lineage.head(1)

now diving this dataframe into 10 csvs and fasta files to run 10 SIMBA jobs in parallel


num_seq_per_batch=500 #how many seqs per fasta file
for i in range(0,ref50_lineage.shape[0],num_seq_per_batch):
    start=i
    end=i+ num_seq_per_batch
    batch_df = ref50_lineage.iloc[start:end]
    batch_df.to_csv(mainDir4 + 'fasta_files/batch_' + str(i) + '.csv')
    
    #Opening a fasta file to write random sequences to
    f= open(mainDir4 + 'fasta_files/batch_' + str(i) + '.fasta', 'w')
    for i in list(range(0,batch_df.shape[0])):
        header = str(batch_df['taxid'].iat[i]) + '_'+ str(batch_df['id2'].iat[i]) + '_' + batch_df['species'].iat[i]
        header = header.replace(' ', '_')
        sequence = batch_df['rand_seq'].iat[i]
        f.write('>' + header + '\n' + sequence + '\n')
    f.close()
